[#1052] adaption to LogicalGraph refactoring
