+++ gradoop-flink/src/test/java/org/gradoop/flink/algorithms/gelly/randomjump/RandomJumpBaseTest.java
@@ -19,7 +19,7 @@ import org.apache.flink.api.java.io.LocalCollectionOutputFormat;
 import org.gradoop.common.model.impl.pojo.Edge;
 import org.gradoop.common.model.impl.pojo.Vertex;
 import org.gradoop.flink.model.GradoopFlinkTestBase;
-import org.gradoop.flink.model.api.epgm.LogicalGraph;
+import org.gradoop.flink.model.impl.epgm.LogicalGraph;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
